update issue 38
Status: In Progress

Added the C# code to class Environment3 for generating the Terrain, and the skybox, and applying effects such as lighting to them, but didn't add the effect files yet. Also I commented all of my code to prevent errors caused by missing effect files and from using game elements while the class is not a game class.