update issue 16 Status: In Progress
removed some useless comments(they were reminders for me to discuss stuff with my team during a meeting) and modified the commented lines to work according to what was discussed in the meeting today. Method will be uncommented when all the methods it uses are implemented.

update issue 98
Status: New
modified names according to last renames