update issue 38
Status: In Progress

Removed the moving first person camera from Environment3 and used a Target Camera from the common classes, because it needs to be unified with the cameras of the hole and the ball.
Removed some incorrect random symbols that were written in Environment3 and were causing errors.
By C1 and C2.